+++ src/rekor/models/consistency_proof.rs
@@ -31,9 +31,9 @@ impl ConsistencyProof {
     /// Verify this consistency proof against the given parameters.
     pub fn verify(
         &self,
-        old_size: usize,
+        old_size: u64,
         old_root: &str,
-        new_size: usize,
+        new_size: u64,
     ) -> Result<(), SigstoreError> {
         // decode hashes from hex and convert them to the required data structure
         // immediately return an error when conversion fails
@@ -47,8 +47,8 @@ impl ConsistencyProof {
         let new_root = hex_to_hash_output(&self.root_hash)?;
 
         Rfc6269Default::verify_consistency(
-            old_size as u64,
-            new_size as u64,
+            old_size,
+            new_size,
             &proof_hashes,
             &old_root,
             &new_root,
